1.add BaseLazyFragment;2.add global config in manifest.
