+++ CommonSDK/src/main/AndroidManifest.xml
@@ -1,2 +1,10 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.vea.atoms.mvp.commonsdk" />
+          package="com.vea.atoms.mvp.commonsdk">
+    <application>
+
+        <meta-data
+                android:name="com.vea.atoms.mvp.commonsdk.core.GlobalConfiguration"
+                android:value="ConfigModule"/>
+
+    </application>
+</manifest>
