+++ Module-gank/src/main/java/com/vea/atoms/mvp/gank/mvp/ui/fragment/GirlFragment.java
@@ -67,11 +67,6 @@ public class GirlFragment extends BaseFragment<GirlPresenter> implements GirlCon
     private IBaseShowItemList<GankItemBean> baseShowItemList;
 
     @Override
-    protected void onUserVisible() {
-
-    }
-
-    @Override
     protected void setupFragmentComponent(AppComponent appComponent) {
         
         DaggerGankComponent
